requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
httpx>=0.28.0
python-dotenv>=1.0.0
scikit-learn>=1.5.0
numpy>=2.0.0
google-generativeai>=0.8.0
jinja2>=3.1.0
mcp>=1.2.0
